@@ -1,7 +1,7 @@
 import Foundation
 
-public struct EmailAddress: Codable {
-    /// format: email
+public struct EmailAddress: Codable, Sendable {
+    /// The email address of the person to whom you are sending an email.
     public var email: String
     
     /// The name of the person to whom you are sending an email.
