@@ -1,11 +1,12 @@
 import Foundation
 
-public struct EmailAttachment: Codable {
-    
+public struct EmailAttachment: Codable, Sendable {
     /// The Base64 encoded content of the attachment.
     public var content: String
     
-    /// The mime type of the content you are attaching. For example, “text/plain” or “text/html”.
+    /// The MIME type of the content you are attaching.
+    /// 
+    /// For example, `“text/plain”` or `“text/html”`.
     public var type: String?
     
     /// The filename of the attachment.
@@ -14,7 +15,10 @@ public struct EmailAttachment: Codable {
     /// The content-disposition of the attachment specifying how you would like the attachment to be displayed.
     public var disposition: String?
     
-    /// The content id for the attachment. This is used when the disposition is set to “inline” and the attachment is an image, allowing the file to be displayed within the body of your email.
+    /// The content ID for the attachment.
+    /// 
+    /// This is used when the disposition is set to “inline” and the attachment is an image,
+    /// allowing the file to be displayed within the body of your email.
     public var contentId: String?
     
     public init(
@@ -38,5 +42,4 @@ public struct EmailAttachment: Codable {
         case disposition
         case contentId = "content_id"
     }
-    
 }
