@@ -1,11 +1,14 @@
 import Foundation
 
-public struct EmailContent: Codable {
-
-    /// The mime type of the content you are including in your email. For example, “text/plain” or “text/html”.
+public struct EmailContent: Codable, Sendable {
+    /// The MIME type of the content you are including in your email.
+    /// 
+    /// For example, `“text/plain”` or `“text/html”`.
     public var type: String
 
-    /// The actual content of the specified mime type that you are including in your email. minLength 1
+    /// The actual content of the specified MIME type that you are including in your email.
+    /// 
+    /// > Important: The minimum length is 1.
     public var value: String
 
     public init(
