@@ -1,10 +1,10 @@
 import Foundation
 
-public struct SendGridError: Error, Decodable {
+public struct SendGridError: Error, Decodable, Sendable {
     public var errors: [SendGridErrorResponse]?
 }
 
-public struct SendGridErrorResponse: Decodable {
+public struct SendGridErrorResponse: Decodable, Sendable {
     public var message: String?
     public var field: String?
     public var help: String?
